Put functions in separate modules. Struct masterRecord and type Data is now in utils.h
