+++ project/include/blackRecord.h
@@ -1,0 +1,6 @@
+#ifndef HOMEWORK_BLACKRECORD_H
+#define HOMEWORK_BLACKRECORD_H
+
+void blackRecord(FILE *ofPTR, FILE *ofPTR_2, FILE *blackrecord, Data client_data, Data transfer);
+
+#endif //HOMEWORK_BLACKRECORD_H
