+++ project/include/masterWrite.h
@@ -1,0 +1,6 @@
+#ifndef HOMEWORK_MASTERWRITE_H
+#define HOMEWORK_MASTERWRITE_H
+
+void masterWrite(FILE *ofPTR, Data Client);
+
+#endif //HOMEWORK_MASTERWRITE_H
