+++ project/include/transactionWrite.h
@@ -1,0 +1,6 @@
+#ifndef HOMEWORK_TRANSACTIONWRITE_H
+#define HOMEWORK_TRANSACTIONWRITE_H
+
+void transactionWrite(FILE *ofPtr, Data transfer);
+
+#endif  // HOMEWORK_TRANSACTIONWRITE_H
